Major refactor of config handling, protocol handling to be far cleaner. More commenting.
Config and protocol handlers have been removed in favour of using the function within the core.py of both systems and converting the resulting dictionary to an object. More work could be done to add functions to check if more than one key exists in a part of the object. Should do this later. Capabilities have been redone inline with the rest of the main.py inits. More commenting and cleanups.